% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_feature_dataset.R
\name{build_feature_dataset}
\alias{build_feature_dataset}
\title{Build feature set for machine learning model}
\usage{
build_feature_dataset(data, mol)
}
\arguments{
\item{data}{combination of MoA and treatments}

\item{mol}{tibble Molecular signatures from untreated sample}
}
\description{
description
}
