% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsible_box.R
\name{collapsible_box_server}
\alias{collapsible_box_server}
\title{Server component for a collapsible HTML box}
\usage{
collapsible_box_server(id, collapsible_title = "Test")
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{collapsible_title}{A character that is used as a title for the box}
}
\description{
Server component for a collapsible HTML box
}
