% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsible_box.R
\name{collapsible_box_ui}
\alias{collapsible_box_ui}
\title{UI Component of collabsible box}
\usage{
collapsible_box_ui(id, ui)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
serve function.}

\item{ui}{Shiny ui components to be shown in the collapsible box.}
}
\description{
UI Component of collabsible box
}
