% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_feature_categories.R
\name{create_feature_categories}
\alias{create_feature_categories}
\title{Map feature names to categories as used in the publication}
\usage{
create_feature_categories(view)
}
\arguments{
\item{view}{A character name of the fiew listing all available features}
}
\description{
Map feature names to categories as used in the publication
}
