% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.combn_pair}
\alias{.combn_pair}
\title{Create combination of a vector of size m}
\usage{
.combn_pair(x, col, m = 2)
}
\arguments{
\item{x}{A data frame}

\item{col}{A character column name of the data frame x}

\item{m}{An integer for the number of elements}
}
\value{
A data frame with all combinations with m columns
}
\description{
Create combination of a vector of size m
}
