% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_synergy_ddr_app.R
\name{launch_synergy_ddr_app}
\alias{launch_synergy_ddr_app}
\title{Launcher for}
\usage{
launch_synergy_ddr_app(fst_folder = NULL)
}
\arguments{
\item{fst_folder}{path to fst files (only for development required)}
}
\value{
Starts a shiny app
}
\description{
Launcher for
}
\examples{
\dontrun{
launch_synergy_ddr_app()
}
}
