% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fst.R
\name{load_fst}
\alias{load_fst}
\title{Wrapper for indexed fst access}
\usage{
load_fst(
  fst_name,
  cols = NULL,
  filter = NULL,
  distinct = FALSE,
  order = NULL,
  stringAsFactors = FALSE,
  peek = FALSE,
  folder = NULL
)
}
\arguments{
\item{fst_name}{A character with the fst file name (optionally the same
location contains a corresponding index file).}

\item{cols}{A character vector of columns to be loaded}

\item{filter}{A named list with column name, value pairs (used for filtering)}

\item{distinct}{A logical indicating if only unique values should be returned}

\item{order}{A character vector, listing columns by that the returned tibble
should be sorted}

\item{stringAsFactors}{A logical, if characters should be cast to facttors}

\item{peek}{A logical, if only the top 10 rows should be returned}

\item{folder}{A character with the path to the fst file}
}
\value{
A tibble listing entries for the defined parameters
}
\description{
This method uses an fst file  (file.fst) and a corresponding index file
(file_index.fst) to optimize the access of fst files. The index file
should contain a column with the name of the column that is indexed in the
fst file and to columns (start, end) that list the first and last value of
an value in the column. The fst file must be sorted by the column.
}
