% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_plot.R
\name{nested_plot_server}
\alias{nested_plot_server}
\title{Server component for displaying an interactive plot that is clickable}
\usage{
nested_plot_server(
  id,
  top_level_name,
  plot_title = NULL,
  summary,
  ignore_click = c("")
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{top_level_name}{A character naming the name of the x cvategories for
the initial plot (used as x title and for the previous button)}

\item{plot_title}{A optional character for the plot title}

\item{summary}{An shiny reactive with the details for the plot output}

\item{ignore_click}{A character vector of x values that should be ignored
when clicked on}
}
\description{
Server component for displaying an interactive plot that is clickable
}
