% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_plot.R
\name{nested_plot_ui}
\alias{nested_plot_ui}
\title{UI component for displaying an interactive plot that is clickable}
\usage{
nested_plot_ui(
  id,
  x_breaks_name = "Breaks",
  info_text = shiny::HTML(""),
  info_title = shiny::h2("")
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
server function.}

\item{x_breaks_name}{A character for the label of an input field for selecting
above the breaks of the x axis of the plot}

\item{info_text}{Small description of the plot and its content}

\item{info_title}{Small title of the plot (wrapped in \link[shiny]{h2})}
}
\description{
UI component for displaying an interactive plot that is clickable
}
