% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_combination_effect.R
\name{predict_combination_effect}
\alias{predict_combination_effect}
\title{Predict treatment combination effect}
\usage{
predict_combination_effect(
  X,
  pred_target,
  predcontrib = FALSE,
  path = NULL,
  threads = 1
)
}
\arguments{
\item{X}{A matrix of all features as used for training}

\item{pred_target}{A character for the score to be predicted ('aoc' or bliss)}

\item{predcontrib}{A logical for returning pe-feature contributions or not}

\item{path}{A character of the path to model files}

\item{threads}{The number of threads to be used by \link{lightgbm}
(name: '{pred_target}_{cnt}.model.txt')}
}
\description{
Predict treatment combination effect
}
