% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_optimal_drug_combination.R
\name{predict_optimal_drug_combination}
\alias{predict_optimal_drug_combination}
\title{Predict best drug combinations based on patient's molecular readouts}
\usage{
predict_optimal_drug_combination(
  mol_df,
  predcontrib = FALSE,
  threads = 1,
  model_path = NULL,
  chunk_size = 50
)
}
\arguments{
\item{mol_df}{A data.frame with molecular readouts from patient}

\item{predcontrib}{A logical if per-feature contributions should be calculated}

\item{threads}{An integer for the number of threads used for parallelization}

\item{model_path}{A character to the path containing model files
(name: '{pred_target}_{cnt}.model.txt')}

\item{chunk_size}{An integer defining the chunck size that is}
}
\value{
best drug combinations based on prediction
}
\description{
Predict best drug combinations based on patient's molecular readouts
}
