% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_prediction.R
\name{prediction_server}
\alias{prediction_server}
\title{Server component for displaying a plot and table output for synergy and
efficacy scores}
\usage{
prediction_server(id, view = list())
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
server function.}

\item{view}{A character of the fst file listing predictions}
}
\description{
Server component for displaying a plot and table output for synergy and
efficacy scores
}
