% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_prediction.R
\name{prediction_ui}
\alias{prediction_ui}
\title{UI component for displaying a plot and table output for synergy and efficacy
scores}
\usage{
prediction_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
server function.}
}
\description{
UI component for displaying a plot and table output for synergy and efficacy
scores
}
