% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_score_data.R
\name{score_data_server}
\alias{score_data_server}
\title{Server component for displaying a plot and table output for synergy and
efficacy scores for an uploaded file}
\usage{
score_data_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
server function.}
}
\description{
Server component for displaying a plot and table output for synergy and
efficacy scores for an uploaded file
}
