% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_score_data.R
\name{score_data_ui}
\alias{score_data_ui}
\title{UI component for displaying a plot and table output for synergy and efficacy
scores of an uploaded file}
\usage{
score_data_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
server function.}
}
\description{
UI component for displaying a plot and table output for synergy and efficacy
scores of an uploaded file
}
