% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_input_values.R
\name{summarize_input_values}
\alias{summarize_input_values}
\title{Summary function for input values}
\usage{
summarize_input_values(feature_annotation, view, active_x)
}
\arguments{
\item{feature_annotation}{A data.frame defining categories of features
(see \code{\link[=create_feature_categories]{create_feature_categories()}})}

\item{view}{A character name of the view that contains input values}

\item{active_x}{A character feature name for that the input statistics should
be summarized}
}
\description{
This function summarizes the input values of the test data for a defined
set of features
}
