% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_shap_values.R
\name{summarize_shap_values}
\alias{summarize_shap_values}
\title{Summary function for SHAP values}
\usage{
summarize_shap_values(
  feature_annotation,
  shap_view,
  active_x,
  readout,
  input_view
)
}
\arguments{
\item{feature_annotation}{data.frame defining categories of features (see
\code{\link[=create_feature_categories]{create_feature_categories()}})}

\item{shap_view}{A character name of the fst file that contains SHAP values}

\item{active_x}{feature name}

\item{readout}{A character of the score to be predicted ('bliss' or 'aoc')}

\item{input_view}{A character name of the fst file that contains feature values}
}
\description{
This function summarizes the SHAP values of the test data for a defined
set of features
}
